+++ code/isa/src/utils.cpp
@@ -11,25 +11,25 @@
 
 using namespace std;
 
-Array<double, 1, Dynamic> logsumexp(const ArrayXXd& array) {
-	Array<double, 1, Dynamic> arrayMax = array.colwise().maxCoeff() - 1.;
+Array<float, 1, Dynamic> logsumexp(const ArrayXXf& array) {
+	Array<float, 1, Dynamic> arrayMax = array.colwise().maxCoeff() - 1.;
 	return arrayMax + (array.rowwise() - arrayMax).exp().colwise().sum().log();
 }
 
 
 
-Array<double, 1, Dynamic> logmeanexp(const ArrayXXd& array) {
-	Array<double, 1, Dynamic> arrayMax = array.colwise().maxCoeff() - 1.;
+Array<float, 1, Dynamic> logmeanexp(const ArrayXXf& array) {
+	Array<float, 1, Dynamic> arrayMax = array.colwise().maxCoeff() - 1.;
 	return arrayMax + (array.rowwise() - arrayMax).exp().colwise().mean().log();
 }
 
 
 
 #ifdef __GXX_EXPERIMENTAL_CXX0X__
-ArrayXXd sampleNormal(int m, int n) {
+ArrayXXf sampleNormal(int m, int n) {
 	mt19937 gen(rand());
-	normal_distribution<double> normal;
-	ArrayXXd samples(m, n);
+	normal_distribution<float> normal;
+	ArrayXXf samples(m, n);
 
 	for(int i = 0; i < samples.size(); ++i)
 		samples(i) = normal(gen);
@@ -38,9 +38,9 @@ ArrayXXd sampleNormal(int m, int n) {
 }
 #else
 #warning "No C++11 support. Using my own implementation of the Box-Muller transform."
-ArrayXXd sampleNormal(int m, int n) {
-	ArrayXXd U1 = ArrayXXd::Random(m, n).abs();
-	ArrayXXd U2 = ArrayXXd::Random(m, n).abs();
+ArrayXXf sampleNormal(int m, int n) {
+	ArrayXXf U1 = ArrayXXf::Random(m, n).abs();
+	ArrayXXf U2 = ArrayXXf::Random(m, n).abs();
 	// Box-Muller transform
 	return (-2. * U1.log()).sqrt() * (2. * PI * U2).cos();
 }
@@ -48,27 +48,27 @@ ArrayXXd sampleNormal(int m, int n) {
 
 
 
-ArrayXXd sampleGamma(int m, int n, int k) {
-	ArrayXXd samples = ArrayXXd::Zero(m, n);
+ArrayXXf sampleGamma(int m, int n, int k) {
+	ArrayXXf samples = ArrayXXf::Zero(m, n);
 
 	for(int i = 0; i < k; ++i)
-		samples -= ArrayXXd::Random(m, n).abs().log();
+		samples -= ArrayXXf::Random(m, n).abs().log();
 
 	return samples;
 }
 
 
 
-VectorXi argsort(const VectorXd& data) {
+VectorXi argsort(const VectorXf& data) {
 	// create pairs of values and indices
-	vector<pair<double, int> > pairs(data.size());
+	vector<pair<float, int> > pairs(data.size());
 	for(int i = 0; i < data.size(); ++i) {
 		pairs[i].first = data[i];
 		pairs[i].second = i;
 	}
 
 	// sort values in descending order
-	sort(pairs.begin(), pairs.end(), greater<pair<double, int> >());
+	sort(pairs.begin(), pairs.end(), greater<pair<float, int> >());
 
 	// store indices
 	VectorXi indices(data.size());
@@ -80,35 +80,35 @@ VectorXi argsort(const VectorXd& data) {
 
 
 
-MatrixXd covariance(const MatrixXd& data) {
-	MatrixXd data_centered = data.colwise() - data.rowwise().mean().eval();
+MatrixXf covariance(const MatrixXf& data) {
+	MatrixXf data_centered = data.colwise() - data.rowwise().mean().eval();
 	return data_centered * data_centered.transpose() / data.cols();
 }
 
 
 
-MatrixXd corrcoef(const MatrixXd& data) {
-	MatrixXd C = covariance(data);
-	VectorXd c = C.diagonal();
+MatrixXf corrcoef(const MatrixXf& data) {
+	MatrixXf C = covariance(data);
+	VectorXf c = C.diagonal();
 	return C.array() / (c * c.transpose()).array().sqrt();
 }
 
 
 
-MatrixXd normalize(const MatrixXd& matrix) {
+MatrixXf normalize(const MatrixXf& matrix) {
 	return matrix.array().rowwise() / matrix.colwise().norm().eval().array();
 }
 
 
 
-double logDetPD(const MatrixXd& matrix) {
+float logDetPD(const MatrixXf& matrix) {
 	return 2. * matrix.llt().matrixLLT().diagonal().array().log().sum();
 }
 
 
 
-MatrixXd deleteRows(const MatrixXd& matrix, vector<int> indices) {
-	MatrixXd result = ArrayXXd::Zero(matrix.rows() - indices.size(), matrix.cols());
+MatrixXf deleteRows(const MatrixXf& matrix, vector<int> indices) {
+	MatrixXf result = ArrayXXf::Zero(matrix.rows() - indices.size(), matrix.cols());
 
 	sort(indices.begin(), indices.end());
 
@@ -127,8 +127,8 @@ MatrixXd deleteRows(const MatrixXd& matrix, vector<int> indices) {
 
 
 
-MatrixXd deleteCols(const MatrixXd& matrix, vector<int> indices) {
-	MatrixXd result = ArrayXXd::Zero(matrix.rows(), matrix.cols() - indices.size());
+MatrixXf deleteCols(const MatrixXf& matrix, vector<int> indices) {
+	MatrixXf result = ArrayXXf::Zero(matrix.rows(), matrix.cols() - indices.size());
 
 	sort(indices.begin(), indices.end());
